requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
tqdm>=4.60.0
wandb>=0.12.0
laplace-torch>=0.1.2 